Renamed AddFriendFrom to FriendAddForm, refactored the form handlers
